Moves initialization structures to a separate folder
