+++ turnip-engine/src/Graphics/Vulkan/Initialization/Instance.h
@@ -2,7 +2,7 @@
 #include <optional>
 #include <vulkan/vulkan.hpp>
 
-#include "Constants.h"
+#include "Graphics/Vulkan/Constants.h"
 #include "Core/Defines.h"
 #include "Core/Logger/Logger.h"
