+++ turnip-engine/src/Graphics/Vulkan/Initialization/VulkanInitializer.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "Common.h"
-#include "BackendVulkan.h"
+#include "Graphics/Vulkan/BackendVulkan.h"
 
 #include "Instance.h"
 #include "PhysicalDevice.h"
