+++ turnip-engine/turnip-engine.vcxproj
@@ -136,13 +136,13 @@
     <ClInclude Include="src\Graphics\OpenGL\BackendOpenGL.h" />
     <ClInclude Include="src\Graphics\Vulkan\BackendVulkan.h" />
     <ClInclude Include="src\Graphics\Vulkan\Constants.h" />
-    <ClInclude Include="src\Graphics\Vulkan\Frame.h" />
-    <ClInclude Include="src\Graphics\Vulkan\Instance.h" />
-    <ClInclude Include="src\Graphics\Vulkan\LogicalDevice.h" />
-    <ClInclude Include="src\Graphics\Vulkan\PhysicalDevice.h" />
-    <ClInclude Include="src\Graphics\Vulkan\Queues.h" />
-    <ClInclude Include="src\Graphics\Vulkan\Swapchain.h" />
-    <ClInclude Include="src\Graphics\Vulkan\VulkanInitializer.h" />
+    <ClInclude Include="src\Graphics\Vulkan\Initialization\Frame.h" />
+    <ClInclude Include="src\Graphics\Vulkan\Initialization\Instance.h" />
+    <ClInclude Include="src\Graphics\Vulkan\Initialization\LogicalDevice.h" />
+    <ClInclude Include="src\Graphics\Vulkan\Initialization\PhysicalDevice.h" />
+    <ClInclude Include="src\Graphics\Vulkan\Initialization\Queues.h" />
+    <ClInclude Include="src\Graphics\Vulkan\Initialization\Swapchain.h" />
+    <ClInclude Include="src\Graphics\Vulkan\Initialization\VulkanInitializer.h" />
     <ClInclude Include="src\Platform\GLFW\EventDebug_GLFW.h" />
     <ClInclude Include="src\Platform\GLFW\Events_GLFW.h" />
     <ClInclude Include="src\Platform\GLFW\Monitor_GLFW.h" />
@@ -161,7 +161,7 @@
     <ClCompile Include="src\Graphics\GraphicsBackend.cpp" />
     <ClCompile Include="src\Graphics\OpenGL\BackendOpenGL.cpp" />
     <ClCompile Include="src\Graphics\Vulkan\BackendVulkan.cpp" />
-    <ClCompile Include="src\Graphics\Vulkan\PhysicalDevice.cpp" />
+    <ClCompile Include="src\Graphics\Vulkan\Initialization\PhysicalDevice.cpp" />
     <ClCompile Include="src\Platform\GLFW\Window_GLFW.cpp" />
     <ClCompile Include="src\Util\CPUInfo\CPUInfo.cpp" />
     <ClCompile Include="src\pch.cpp">
