+++ turnip-engine/turnip-engine.vcxproj.filters
@@ -176,22 +176,22 @@
     <ClInclude Include="src\Graphics\Vulkan\Constants.h">
       <Filter>Graphics\Vulkan</Filter>
     </ClInclude>
-    <ClInclude Include="src\Graphics\Vulkan\Instance.h">
+    <ClInclude Include="src\Graphics\Vulkan\Initialization\Instance.h">
       <Filter>Graphics\Vulkan</Filter>
     </ClInclude>
-    <ClInclude Include="src\Graphics\Vulkan\LogicalDevice.h">
+    <ClInclude Include="src\Graphics\Vulkan\Initialization\LogicalDevice.h">
       <Filter>Graphics\Vulkan</Filter>
     </ClInclude>
-    <ClInclude Include="src\Graphics\Vulkan\PhysicalDevice.h">
+    <ClInclude Include="src\Graphics\Vulkan\Initialization\PhysicalDevice.h">
       <Filter>Graphics\Vulkan</Filter>
     </ClInclude>
-    <ClInclude Include="src\Graphics\Vulkan\Queues.h">
+    <ClInclude Include="src\Graphics\Vulkan\Initialization\Queues.h">
       <Filter>Graphics\Vulkan</Filter>
     </ClInclude>
-    <ClInclude Include="src\Graphics\Vulkan\Swapchain.h">
+    <ClInclude Include="src\Graphics\Vulkan\Initialization\Swapchain.h">
       <Filter>Graphics\Vulkan</Filter>
     </ClInclude>
-    <ClInclude Include="src\Graphics\Vulkan\VulkanInitializer.h">
+    <ClInclude Include="src\Graphics\Vulkan\Initialization\VulkanInitializer.h">
       <Filter>Graphics\Vulkan</Filter>
     </ClInclude>
     <ClInclude Include="src\Platform\GLFW\EventDebug_GLFW.h">
@@ -223,7 +223,7 @@
       <Filter>Util\CPUInfo</Filter>
     </ClInclude>
     <ClInclude Include="src\pch.h" />
-    <ClInclude Include="src\Graphics\Vulkan\Frame.h" />
+    <ClInclude Include="src\Graphics\Vulkan\Initialization\Frame.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\Core\Engine\TurnipEngine.cpp">
@@ -241,7 +241,7 @@
     <ClCompile Include="src\Graphics\Vulkan\BackendVulkan.cpp">
       <Filter>Graphics\Vulkan</Filter>
     </ClCompile>
-    <ClCompile Include="src\Graphics\Vulkan\PhysicalDevice.cpp">
+    <ClCompile Include="src\Graphics\Vulkan\Initialization\PhysicalDevice.cpp">
       <Filter>Graphics\Vulkan</Filter>
     </ClCompile>
     <ClCompile Include="src\Platform\GLFW\Window_GLFW.cpp">
